+++ src/main/java/org/folio/dew/batch/bursarfeesfines/AccountItemProcessor.java
@@ -4,7 +4,7 @@ import java.util.Collections;
 import java.util.Map;
 import org.folio.dew.domain.dto.Account;
 import org.folio.dew.domain.dto.bursarfeesfines.BursarFormat;
-import org.folio.dew.utils.FormatterUtils;
+import org.folio.dew.utils.BursarFeesFinesUtils;
 import org.springframework.batch.core.StepExecution;
 import org.springframework.batch.core.annotation.BeforeStep;
 import org.springframework.batch.core.configuration.annotation.StepScope;
@@ -20,14 +20,14 @@ public class AccountItemProcessor implements ItemProcessor<Account, BursarFormat
   @Override
   public BursarFormat process(Account item) {
     BursarFormat format = new BursarFormat();
-    format.setEmployeeId(FormatterUtils.getEmployeeId(item.getUserId(), userIdMap));
-    format.setAmount(FormatterUtils.normalizeAmount(item.getAmount()));
+    format.setEmployeeId(BursarFeesFinesUtils.getEmployeeId(item.getUserId(), userIdMap));
+    format.setAmount(BursarFeesFinesUtils.normalizeAmount(item.getAmount()));
     format.setTransactionDate(
-        FormatterUtils.getTransactionDate(item.getMetadata().getCreatedDate()));
+        BursarFeesFinesUtils.getTransactionDate(item.getMetadata().getCreatedDate()));
     format.setSfs("SFS");
     format.setTermValue("    ");
-    format.setDescription(FormatterUtils.getItemTypeDescription(item.getFeeFineType()));
-    format.setItemType(FormatterUtils.getItemType());
+    format.setDescription(BursarFeesFinesUtils.getItemTypeDescription(item.getFeeFineType()));
+    format.setItemType(BursarFeesFinesUtils.getItemType());
     return format;
   }
